:art: Removed dashboard state and created selector
