+++ src/state/ducks/index.ts
@@ -4,15 +4,11 @@ import { combineReducers } from 'redux';
 import { all, fork } from 'redux-saga/effects';
 
 import studentReducer from './student/reducer';
-import dashboardReducer from './dashboard/reducer';
 import { studentSaga } from './student/saga';
 import { IStudentState } from './student/types';
-import { IDashboardState } from './dashboard/types';
-import { dashboardSaga } from './dashboard/saga';
 
 export interface IApplicationState {
   student: IStudentState;
-  dashboard: IDashboardState;
 }
 
 export const useAppSelector: TypedUseSelectorHook<IApplicationState> =
@@ -20,9 +16,8 @@ export const useAppSelector: TypedUseSelectorHook<IApplicationState> =
 
 export const rootReducer = combineReducers<IApplicationState>({
   student: studentReducer,
-  dashboard: dashboardReducer,
 });
 
 export function* rootSaga() {
-  yield all([fork(studentSaga), fork(dashboardSaga)]);
+  yield all([fork(studentSaga)]);
 }
